Move Response class to NuRadioReco/detector/response.py
