+++ NuRadioReco/detector/RNO_G/rnog_detector.py
@@ -1,11 +1,10 @@
 from NuRadioReco.detector.RNO_G.db_mongo_read import Database, _convert_astro_time_to_datetime
-import NuRadioReco.framework.base_trace
+from NuRadioReco.detector.response import Response
 from NuRadioReco.utilities import units
 
-from scipy import interpolate
 from functools import wraps
-import astropy.time
 
+import astropy.time
 import datetime
 import numpy as np
 import collections
@@ -778,7 +777,7 @@ class Detector():
     @_check_detector_time
     def get_signal_chain_response(self, station_id, channel_id):
         """
-        Returns a `rnog_detector.Response` object which describes the complex response of the
+        Returns a `detector.response.Response` object which describes the complex response of the
         entire signal chain, i.e., the combined reponse of all components of one
         channel. For example: IGLU, fiber-cable, DRAB, coax-cable, RADIANT.
 
@@ -794,7 +793,7 @@ class Detector():
         Returns
         -------
 
-        response: rnog_detector.Response
+        response: `detector.response.Response`
             Returns combined response of the channel
         """
         signal_chain_dict = self.get_channel_signal_chain(
@@ -870,7 +869,7 @@ class Detector():
 
         signal_chain_components: dict
             A dictionaries with the keys being the names of the components and the values their weights in
-            the signal chain. (For an explanation of a weight see `rnog_detector.Response`)
+            the signal chain. (For an explanation of a weight see `detector.response.Response`)
         """
         signal_chain_dict = self.get_channel_signal_chain(
             station_id, channel_id)
@@ -1219,344 +1218,6 @@ class Detector():
         self.__set_detector_time(time)
 
 
-class Response:
-    """
-    This class provides an interface to read-in and apply the complex response functions of the
-    various components of the signal chain of a RNO-G channel. The class allows to combine the
-    response of several components into one response by multiplying them.
-
-    Examples
-    --------
-
-    .. code-block::
-        response = det.get_signal_chain_response(station_id=24, channel_id=0)
-
-        # Multipy the response to a trace. The multiply operator takes care of everything
-        trace_at_readout = trace_at_antenna * response
-
-        # getting the complex response as array
-        freq = np.arange(50, 1000) * units.MHz
-        complex_resp = response(freq)
-
-    """
-
-    def __init__(self, frequency, y, y_unit, time_delay=0, weight=1,
-                 name="default", station_id=None, channel_id=None,
-                 remove_time_delay=True, debug_plot=False):
-        """
-        Parameters
-        ----------
-
-        frequency : list(float)
-            The frequency vector at which the response is measured. Unit has to be GHz.
-
-        y : [list(float), list(float)]
-            The measured response. First entry is the vector of the measured amplitude, the second entry is the measured phase.
-            The unit of both entries is specified with the next argument.
-
-        y_unit : [str, str]
-            The first entry specifies the unit of the measured amplitude. Options are "dB", "MAG" and "mag".
-            The second entry specifies the unit of the measured phase. Options are "rad" and "deg".
-
-        time_delay : float (Default: 0)
-            "Average" group delay. Read from database. Will be used to normalize the phase and stored alongside response
-
-        weight : float (Default: 1)
-            Specifies the weight with which this component reponse "adds" to the total signal-chain response or data.
-            Its the exponent of the complex multiplicitive gain. That means that a value of 1 means to linear multiply this
-            reponse while a value of -1 means to divide by this reponse.
-
-        name : str (Default: "default")
-            Give the response a name. This is only use for printing purposes.
-
-        station_id : int (Default: None)
-            Associated station id.
-
-        channel_id : int (Default: None)
-            Associated channel id.
-
-        remove_time_delay : bool (Default: True)
-            If True, remove `time_delay` from response.
-
-        debug_plot : bool (Default: False)
-            If True, produce a debug plot
-        """
-
-        self.__names = [name]
-        self._station_id = station_id
-        self._channel_id = channel_id
-
-        self._sanity_check = True # Tmp
-
-        if self._station_id is None or self._channel_id is None:
-            logging.error(f"Station and channel id were not defined for response {name}. Please do that.")
-
-        self.__frequency = np.array(frequency) * units.GHz
-
-        if y[0] is None or y[1] is None:
-            raise ValueError("Data for response incomplete, detected \"None\"")
-
-        y_ampl, y_phase = np.array(y)
-        if y_unit[0] == "dB":
-            gain = 10 ** (y_ampl / 20)
-        elif y_unit[0].lower() == "mag":
-            gain = y_ampl
-        else:
-            raise KeyError
-
-        if y_unit[1].lower() == "deg":
-            if np.max(np.abs(y_phase)) < 2 * np.pi:
-                logging.warning("Is the phase really in deg? Does not look like it... "
-                                f"Do not convert to rad. Phase: {y_phase}")
-            else:
-                y_phase = np.deg2rad(y_phase)
-        elif y_unit[1].lower() == "rad":
-            y_phase = y_phase
-        else:
-            raise KeyError
-
-
-        if remove_time_delay:
-            y_phase_orig = np.copy(np.unwrap(y_phase))
-            _response = gain * np.exp(1j * (y_phase + 2 * np.pi * time_delay * self.__frequency))
-            y_phase = np.angle(_response)
-        else:
-            time_delay = 0
-
-        y_phase = np.unwrap(y_phase)
-
-        self.__gains = [interpolate.interp1d(
-            self.__frequency, gain, kind="linear", bounds_error=False, fill_value=0)]
-
-        self.__phases = [interpolate.interp1d(
-            self.__frequency, y_phase, kind="linear", bounds_error=False, fill_value=0)]
-
-        self.__weights = [weight]
-        self.__time_delays = [weight * time_delay]
-
-        if debug_plot:
-            from matplotlib import pyplot as plt
-            fig, axs = plt.subplots(3, 1, sharex=True)  #, gridspec_kw=dict(hspace=0.03))
-            axs[0].set_title(name)
-            frequency_interp = np.linspace(self.__frequency[0], self.__frequency[-1], 10000)
-            axs[0].plot(self.__frequency, gain, "C0o", label="data", markersize=2)
-            axs[0].plot(frequency_interp, self.__gains[0](frequency_interp), "C1--", lw=1, label="interpolation")
-
-            if remove_time_delay:
-                axs[1].plot(self.__frequency, y_phase_orig, "C0o", markersize=2, label="Original phase")
-                ax2 = axs[1].twinx()
-                ax2.plot(self.__frequency, y_phase, "C2o", markersize=2, label=f"Excl. time delay of {self.__time_delays[0]:.2f}ns")
-                ax2.plot(frequency_interp, self.__phases[0](frequency_interp), "C3--", lw=1, label="interpolation")
-                ax2.set_ylabel("norm. phase / rad")
-                ax2.legend(fontsize=5)
-            else:
-                axs[1].plot(self.__frequency, y_phase, "C0o", markersize=2)
-                axs[1].plot(frequency_interp, self.__phases[0](frequency_interp), "C1--", lw=1)
-
-            axs[0].set_ylabel("gain")
-            axs[1].set_ylabel("phase / rad")
-
-            axs[0].legend(fontsize=5)
-            # axs[1].legend(fontsize=5)
-
-            group_delay = -np.diff(y_phase) / np.diff(self.__frequency)[0] / (2 * np.pi)
-            axs[2].plot(self.__frequency[:-1], group_delay, "C0o", markersize=2)
-            frequency_mask = np.all([50 * units.MHz < self.__frequency[:-1], self.__frequency[:-1] < 800 * units.MHz], axis=0)
-            axs[2].set_ylim(np.amin(group_delay[frequency_mask]), np.amax(group_delay[frequency_mask]))
-            axs[2].set_ylabel("group delay / ns")
-
-            axs[-1].set_xlabel("frequency / GHz")
-            fig.tight_layout()
-            plt.savefig(f'{name}_{self._station_id}_{self._channel_id}_'
-                        f'{datetime.datetime.utcnow().strftime("%Y%m%dT%H%M%S")}_debug.png', transparent=False)
-            plt.close()
-
-    def __call__(self, freq, component_name=None):
-        """
-        Returns the complex response for a given frequency.
-
-        Parameters
-        ----------
-
-        freq: list(float)
-            The frequencies for which to get the response.
-
-        Returns
-
-        response: np.array(np.complex128)
-            The complex response at the desired frequencies
-        """
-        response = np.ones_like(freq, dtype=np.complex128)
-
-        for gain, phase, weight, name in zip(self.__gains, self.__phases, self.__weights, self.__names):
-
-            if component_name is not None and component_name != name:
-                continue
-
-            _gain = gain(freq / units.GHz)
-
-            # to avoid RunTime warning and NANs in total reponse
-            if weight == -1:
-                _gain = np.where(_gain > 0, _gain, 1e-6)
-
-            response *= (_gain * np.exp(1j * phase(freq / units.GHz))) ** weight
-
-        if np.allclose(response, np.ones_like(freq, dtype=np.complex128)):
-            if component_name is not None:
-                raise ValueError("Returned response is equal to 1. "
-                                f"Did you requested a non-existing component ({component_name})? "
-                                f"Options are: {self.__names}")
-            else:
-                self.logger.warning("Returned response is equal to 1.")
-
-        return response
-
-    def get_names(self):
-        """ Get list of the names of all individual responses """
-        return self.__names
-
-    def __mul__(self, other):
-        """
-        Define multiplication operator for
-            - Other objects of the same class
-            - Objects of type NuRadioReco.framework.base_trace
-
-        TODO: Multiplication with base_trace will just work when the spectrum is one dimensional. Check if
-        there is a case where the spectrum could be 3 (or more) dimensional and adjust the class
-
-        """
-
-        if isinstance(other, Response):
-            if self._station_id != other._station_id or \
-                self._channel_id != other._channel_id:
-                logging.error("It looks like you are combining responses from "
-                                  f"two different channels: {self._station_id}.{self._channel_id} "
-                                  f" vs {other._station_id}.{other._channel_id} (station_id.channel_id)")
-            # Store each response individually: append/concatenate lists of gains and phases.
-            # The multiplication happens in __call__.
-            self.__names += other.__names
-            self.__gains += other.__gains
-            self.__phases += other.__phases
-            self.__weights += other.__weights
-            self.__time_delays += other.__time_delays
-            return self
-
-        elif isinstance(other, NuRadioReco.framework.base_trace.BaseTrace):
-
-            if self._sanity_check:
-                trace_length = other.get_number_of_samples() / other.get_sampling_rate()
-                time_delay = self._get_time_delay()
-                if time_delay > trace_length / 2:
-                    logging.warning("The time shift appiled by the response is larger than half the trace length:\n\t"
-                                    f"{time_delay:.2f} vs {trace_length:.2f}")
-
-            spec = other.get_frequency_spectrum()
-            freqs = other.get_frequencies()
-            spec *= self(freqs)  # __call__
-            other.add_trace_start_time(np.sum(self.__time_delays))
-            other.set_frequency_spectrum(spec, sampling_rate="same")
-            return other
-
-        elif isinstance(other, np.ndarray):
-            raise TypeError("You try to multiply a `Response` object with a numpy array, "
-                            "only `Response` or `BaseTrace` is allowed. "
-                            "Did you call `get_trace()` or `get_frequency_spectrum()` on `BaseTrace`?")
-
-        else:
-            raise TypeError(f"Response multiplied with unknown type: {type(other)}")
-
-    def __rmul__(self, other):
-        """ Same as mul """
-        return self.__mul__(other)
-
-    def __str__(self):
-        ampl = 20 * np.log10(np.abs(self(0.5 * units.GHz)))
-        return "Response of " + ", ".join([f"{name} ({weight})" for name, weight in zip(self.get_names(), self.__weights)]) \
-            + f": |R(0.5 GHz)| = {ampl:.2f} dB (amplitude) ({np.sum(self.__time_delays):.2f} ns)"
-
-    def plot(self, show=False, in_dB=True):
-        import matplotlib.pyplot as plt
-
-        freqs = np.linspace(0, 1.4) * units.GHz
-
-        fig, ax = plt.subplots()
-        for gain, weight, name in zip(self.__gains, self.__weights, self.__names):
-            _gain = gain(freqs)
-
-            ls = "-" if weight == 1 else "--"
-
-            if in_dB:
-                mask = _gain > 0  # to avoid RunTime warning
-                ax.plot(freqs[mask] / units.MHz, 20 * np.log10(_gain[mask]), ls=ls, label=name)
-            else:
-                ax.plot(freqs / units.MHz, _gain, label=name + f": {weight}", ls=ls)
-
-        _gain = np.abs(self(freqs))
-        if in_dB:
-            mask = _gain > 0  # to avoid RunTime warning
-            ax.plot(freqs[mask] / units.MHz, 20 * np.log10(_gain[mask]), color="k", label="total")
-        else:
-            ax.plot(freqs / units.MHz, _gain, color="k", label="total")
-
-        ax.set_xlabel("frequency / MHz")
-        if in_dB:
-            ax.set_ylabel("gain / dB")
-        else:
-            ax.set_ylabel("gain")
-            ax.set_yscale("log")
-
-        ax.legend()
-        ax.grid()
-
-        if show:
-            plt.show()
-        else:
-            return fig, ax
-
-    def get_time_delay(self):
-        """ Get time delay from DB """
-        return np.sum(self.__time_delays)
-
-    def _get_time_delay(self):
-        """
-        Calculate time delay from phase of the stored complex response function.
-        This is not the time delay which is stored in the DB and which is used in
-        the `__init__()` to normalize the reponse function. Rather, its the remaining
-        group delay.
-
-        The time delay is calculated as the mean between 195 and 205 MHz.
-
-        Returns
-        -------
-
-        time_delay1 : float
-            The time delay at ~ 200 MHz
-        """
-
-        freqs = np.arange(0.05, 1.2, 0.001) * units.GHz
-
-        response = self(freqs)
-
-        delta_freq = np.diff(freqs)
-
-        phase = np.angle(response)
-        time_delay = -np.diff(np.unwrap(phase)) / delta_freq / 2 / np.pi
-
-        mask = np.all([195 * units.MHz < freqs, freqs < 250 * units.MHz], axis=0)[:-1]
-        time_delay1 = np.mean(time_delay[mask])
-
-        # fit the unwrapped phase with a linear function
-        popt = np.polyfit(freqs, np.unwrap(phase), 1)
-        time_delay2 = -popt[0] / (2 * np.pi)
-
-        if np.abs(time_delay1 - time_delay2) > 0.1:
-            logging.warning("Calculation of time delay. The two methods yield different results: "
-                            f"{time_delay1:.1f} / {time_delay2:.1f} for {self.get_names()}. Return the former...")
-
-        return time_delay1
-
-
-
 if __name__ == "__main__":
 
     from NuRadioReco.detector import detector
@@ -1565,3 +1226,17 @@ if __name__ == "__main__":
                             database_connection='RNOG_public', select_stations=24)
 
     det.update(datetime.datetime(2023, 8, 2, 0, 0))
+
+
+    response = det.get_signal_chain_response(station_id=24, channel_id=0)
+
+    from NuRadioReco.framework import electric_field
+    ef = electric_field.ElectricField(channel_ids=[0])
+    ef.set_frequency_spectrum(np.ones(1025, dtype=complex), sampling_rate=2.4)
+
+    # Multipy the response to a trace. The multiply operator takes care of everything
+    trace_at_readout = ef * response
+
+    # getting the complex response as array
+    freq = np.arange(50, 1000) * units.MHz
+    complex_resp = response(freq)
